+++ components/Common.module.scss
@@ -19,12 +19,7 @@ $nhsBlue: #005eb8;
   color: $nhsBlue;
 }
 
-.cardContainer {
-  border: none;
-}
-
 .panelDiv {
-  background-color: #f0f4f5;
   padding: 0.75rem;
   border-radius: 6px;
 }
