+++ components/home/Home.Module.scss
@@ -1,0 +1,5 @@
+@import "../Common.module.scss";
+
+.ull {
+  list-style-type: square;
+}
