+++ components/home/HomeHeaderSection.tsx
@@ -1,0 +1,26 @@
+const HomeHeaderSection = () => {
+  return (
+    <section className="nhsuk-hero">
+      <div className="nhsuk-width-container nhsuk-hero--border app-width-container">
+        <div className="nhsuk-grid-row">
+          <div className="nhsuk-grid-column-two-thirds">
+            <div className="nhsuk-hero__wrapper app-hero__wrapper">
+              <h1 className="nhsuk-u-margin-bottom-4">
+                Welcome to TIS Self-Service
+              </h1>
+              <p className="nhsuk-body-l nhsuk-u-margin-bottom-1">
+                This home page is a work-in-progress.
+              </p>
+              <p className="nhsuk-body-m nhsuk-u-margin-bottom-1">
+                This home page is a work-in-progress. This home page is a
+                work-in-progress.
+              </p>
+            </div>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default HomeHeaderSection;
