--- components/profile/personal-details/PersonalDetailsComponent.module.scss
@@ -1,16 +1,0 @@
-@import "../../Common.module.scss";
-
-.personalDetailContainer {
-  margin: 10px 10px 20px 10px;
-  font-size: 16px;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.name {
-  color: $nhsBlue;
-  font-size: 1.5rem;
-  margin-bottom: 0px;
-  font-weight: 500;
-}
